Change tag variable names.
This is necessary to avoid tag clobbering. With the old behavior, each tag
shared the root 'isidore' section of the variable tree. The default behavior of
Ansible is for the last tag alphabetically to override all other tags.
Unfortunately the ability to merge the dictionaries when populating the Ansible
variable tree has been deprecated in a rather questionable decision. This
solution admittedly doesn't look that pretty but it prevents any tags from
being clobbered.