Fix version numbering
Pip doesn't allow package version numbers to contain a -. Additionally it
requires them to end in a number. Add a 0 to the end of the version number to
satisfy the latter requirement and add a sed to lib/Makefile to replace any -
characters with a . for the Python package version number.