Do not sort var print output
This is important as otherwise referencing list elements via their indecies can
result in the wrong item being targeted.

Take the following list:

{ "list": [c, b, a] }

list[0] is c, but when running var print, it was printed as follows:

list:
- a
- b
- c

This would cause a reasonable person to assume list[0] was a, not c. With the
changes introduced in this commit, it now gets printed as expected:

list:
- c
- b
- a