Insert rows into Metadata table if they don't exist
Previously the name and motd rows were updated only if they already
existed. Whereas they're not inserted as part of the initial database
population, this resulted in the initial setting of the name and motd
failing. This employs a more rugged solution that is elegant and allows
for better backwards compability with older installations by not
requiring the database to be repopulated from scratch.